refactor: InfoLabel 컴포넌트를 features -> shared 레이어로 이동
