--- src/features/cart/ui/InfoLabel.style.ts
@@ -1,17 +1,0 @@
-import styled from '@emotion/styled';
-
-export const InfoLabel = styled.div`
-  width: 100%;
-  padding: 10px 0;
-  display: flex;
-  align-items: center;
-  font-size: 12px;
-  font-weight: 500;
-`;
-
-export const InfoIcon = styled.img`
-  width: 16px;
-  height: 16px;
-  margin-right: 6px;
-  vertical-align: middle;
-`;
